refactor: 오탈자 수정
