+++ src/main/kotlin/com/server/dpmcore/cohort/application/CohortQueryService.kt
@@ -4,7 +4,7 @@ import com.server.dpmcore.cohort.application.config.CohortProperties
 import com.server.dpmcore.cohort.domain.exception.CohortNotFoundException
 import com.server.dpmcore.cohort.domain.model.CohortId
 import com.server.dpmcore.cohort.domain.port.inbound.CohortQueryUseCase
-import com.server.dpmcore.cohort.domain.port.outbount.CohortPersistencePort
+import com.server.dpmcore.cohort.domain.port.outbound.CohortPersistencePort
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
