+++ src/main/kotlin/com/server/dpmcore/cohort/infrastructure/repository/CohortRepository.kt
@@ -3,7 +3,7 @@ package com.server.dpmcore.cohort.infrastructure.repository
 import com.linecorp.kotlinjdsl.querydsl.expression.col
 import com.linecorp.kotlinjdsl.spring.data.SpringDataQueryFactory
 import com.server.dpmcore.cohort.domain.model.CohortId
-import com.server.dpmcore.cohort.domain.port.outbount.CohortPersistencePort
+import com.server.dpmcore.cohort.domain.port.outbound.CohortPersistencePort
 import com.server.dpmcore.cohort.infrastructure.entity.CohortEntity
 import com.server.dpmcore.common.jdsl.singleQueryOrNull
 import org.springframework.stereotype.Repository
